+++ lib/src/models/guild/member.dart
@@ -6,6 +6,7 @@ import 'package:nyxx/src/models/permissions.dart';
 import 'package:nyxx/src/models/role.dart';
 import 'package:nyxx/src/models/snowflake.dart';
 import 'package:nyxx/src/models/snowflake_entity/snowflake_entity.dart';
+import 'package:nyxx/src/models/user/avatar_decoration_data.dart';
 import 'package:nyxx/src/models/user/user.dart';
 import 'package:nyxx/src/utils/flags.dart';
 
@@ -61,6 +62,9 @@ class Member extends PartialMember {
   /// The hash of this member's avatar image.
   final String? avatarHash;
 
+  /// The hash of this member's banner image.
+  final String? bannerHash;
+
   /// A list of the IDs of the roles this member has.
   final List<Snowflake> roleIds;
 
@@ -88,6 +92,12 @@ class Member extends PartialMember {
   /// The time until which this member is timed out.
   final DateTime? communicationDisabledUntil;
 
+  /// The member's guild avatar decoration data.
+  final AvatarDecorationData? avatarDecorationData;
+
+  /// The member's guild avatar decoration.
+  final String? avatarDecorationHash;
+
   /// {@macro member}
   /// @nodoc
   Member({
@@ -96,6 +106,7 @@ class Member extends PartialMember {
     required this.user,
     required this.nick,
     required this.avatarHash,
+    required this.bannerHash,
     required this.roleIds,
     required this.joinedAt,
     required this.premiumSince,
@@ -105,6 +116,8 @@ class Member extends PartialMember {
     required this.isPending,
     required this.permissions,
     required this.communicationDisabledUntil,
+    required this.avatarDecorationData,
+    required this.avatarDecorationHash,
   });
 
   /// The roles this member has.
@@ -121,6 +134,23 @@ class Member extends PartialMember {
             ..avatars(),
           hash: avatarHash!,
         );
+
+  CdnAsset? get avatarDecoration => avatarDecorationHash == null
+      ? null
+      : CdnAsset(
+          client: manager.client,
+          base: HttpRoute()..avatarDecorationPresets(),
+          hash: avatarDecorationHash!,
+        );
+
+  CdnAsset get banner => CdnAsset(
+        client: manager.client,
+        base: HttpRoute()
+          ..guilds(id: manager.guildId.toString())
+          ..users(id: id.toString())
+          ..banners(),
+        hash: bannerHash!,
+      );
 }
 
 /// Flags that can be applied to a [Member].
@@ -137,6 +167,23 @@ class MemberFlags extends Flags<MemberFlags> {
   /// This member has started the guild's onboarding process.
   static const startedOnboarding = Flag<MemberFlags>.fromOffset(3);
 
+  /// Member is a guest and can only access the voice channel they were invited to.
+  static const isGuest = Flag<MemberFlags>.fromOffset(4);
+
+  /// Member has started Server Guide new member actions.
+  static const startedHomeActions = Flag<MemberFlags>.fromOffset(5);
+
+  /// Member has completed Server Guide new member actions.
+  static const completedHomeActions = Flag<MemberFlags>.fromOffset(6);
+
+  /// Member's username, display name, or nickname is blocked by AutoMod.
+  static const automodQuarantinedUsername = Flag<MemberFlags>.fromOffset(7);
+
+  // 1 << 8 is AUTOMOD_QUARANTINED_BIO but it's undocumented and deprecated
+
+  /// Member has dismissed the DM settings upsell
+  static const dmSettingsUpsellAcknowledged = Flag<MemberFlags>.fromOffset(9);
+
   /// Whether this member has the [didRejoin] flag.
   bool get hasRejoined => has(didRejoin);
 
@@ -148,6 +195,21 @@ class MemberFlags extends Flags<MemberFlags> {
 
   /// Whether this member has the [startedOnboarding] flag.
   bool get didStartOnboarding => has(startedOnboarding);
+
+  /// Whether this member has the [isGuest] flag.
+  bool get isGuestMember => has(isGuest);
+
+  /// Whether this member has the [startedHomeActions] flag.
+  bool get didStartHomeActions => has(startedHomeActions);
+
+  /// Whether this member has the [completedHomeActions] flag.
+  bool get didCompleteHomeActions => has(completedHomeActions);
+
+  /// Whether this member has the [automodQuarantinedUsername] flag.
+  bool get hasAutomodQuarantinedUsername => has(automodQuarantinedUsername);
+
+  /// Whether this member has the [dmSettingsUpsellAcknowledged] flag.
+  bool get didAcknowledgeDmSettingsUpsell => has(dmSettingsUpsellAcknowledged);
 
   /// Create a new [MemberFlags].
   const MemberFlags(super.value);